+++ pydragonfly/sdk/resources/analysis.py
@@ -1,4 +1,3 @@
-import json
 import dataclasses
 from typing import Optional, List
 from typing_extensions import Literal
@@ -8,6 +7,7 @@ from django_rest_client import (
     APIResource,
     RetrievableAPIResourceMixin,
     ListableAPIResourceMixin,
+    CreateableAPIResourceMixin,
     PaginationAPIResourceMixin,
 )
 from django_rest_client.types import Toid, TParams
@@ -21,12 +21,14 @@ class CreateAnalysisRequestBody:
     root: bool = False
     os: Optional[Literal["WINDOWS", "LINUX"]] = None
     arguments: Optional[List[str]] = None
+    dll_entrypoints: Optional[List[str]] = None
 
 
 class Analysis(
     APIResource,
     RetrievableAPIResourceMixin,
     ListableAPIResourceMixin,
+    CreateableAPIResourceMixin,
     PaginationAPIResourceMixin,
 ):
     """
@@ -53,21 +55,23 @@ class Analysis(
         sample_buffer: bytes,
         params: Optional[TParams] = None,
     ) -> APIResponse:
-        url = "api/create_analysis"
-        post_data = {
-            "data": json.dumps(
-                {k: v for k, v in dataclasses.asdict(data).items() if v is not None}
-            )
-        }
-        post_files = {"sample": (sample_name, sample_buffer)}
-        response = cls._request(
+        # first: POST sample uploading it
+        resp1 = cls._request(
             "POST",
-            url=url,
-            data=post_data,
-            files=post_files,
+            url="api/sample",
+            files={"sample": (sample_name, sample_buffer)},
+        )
+        # second: POST analysis using the new sample ID
+        resp2 = cls._request(
+            "POST",
+            url=cls.class_url(),
+            json={
+                **{k: v for k, v in dataclasses.asdict(data).items() if v is not None},
+                "sample_id": resp1.data["id"],
+            },
             params=params,
         )
-        return response
+        return resp2
 
     @classmethod
     def aggregate_evaluations(